Created changes to fix issues with footer
For the home page it is at the proper bottom of the full page and for the contact page it is at the bottom. The tutoring page didn't have the css page linked to it and I did it for connecting the pages branch, but I linked it here again.